+++ Assets/Scripts/CharacterState/CharacterStates.cs
@@ -1,0 +1,18 @@
+
+public enum CharacterStates
+{
+    Idle,
+    Run,
+    Jump,
+    Crouch,
+    Hurt,
+    Dead,
+    Block,
+
+    // Attacks
+    NuetralAttack,
+    FrontAttack,
+    BackAttack,
+    LowAttack,
+    HighAttack
+}
